backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4

# HTTP & Utilities
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1

# Testing
pytest==8.0.0
httpx>=0.28.1